scripts/requirements/requirements.local.txt
#aliyunsdkcore
aliyun-python-sdk-ecs
alibabacloud_vpc20160428
alibabacloud_ecs20140526
langchain
langchain-community
langchain-core
langsmith
pypdf
faiss-cpu
pyaudio
selenium
tigeropen
hcloud
tavily-python
opencv-python
webdriver-manager
azure-identity
azure-cognitiveservices-speech
google-cloud-speech
google-cloud-texttospeech
google-cloud-storage
matplotlib
pdf2image
lz4
pyperclip
psycopg2
moviepy
